use dcrlibwallet staking PR
